+++ server/src/main/java/nl/tudelft/oopp/group39/building/exceptions/BuildingNotFoundException.java
@@ -1,7 +1,7 @@
 package nl.tudelft.oopp.group39.building.exceptions;
 
 public class BuildingNotFoundException extends RuntimeException {
-    public BuildingNotFoundException(int id) {
+    public BuildingNotFoundException(long id) {
         super("Room with id " + id + " wasn't found.");
     }
 }
