+++ server/src/main/java/nl/tudelft/oopp/group39/building/services/BuildingService.java
@@ -32,7 +32,7 @@ public class BuildingService {
      * @return the Building that was found.
      * @throws BuildingNotFoundException when no building is found.
      */
-    public Building readBuilding(Integer id) throws BuildingNotFoundException {
+    public Building readBuilding(Long id) throws BuildingNotFoundException {
         return buildingRepository.findById(id)
             .orElseThrow(() -> new BuildingNotFoundException(id));
     }
@@ -40,7 +40,7 @@ public class BuildingService {
     /**
      * Doc. TODO Sven
      */
-    public Building deleteBuilding(Integer id) throws BuildingNotFoundException {
+    public Building deleteBuilding(Long id) throws BuildingNotFoundException {
         try {
             Building rf = readBuilding(id);
             buildingRepository.delete(readBuilding(id));
@@ -60,7 +60,7 @@ public class BuildingService {
     /**
      * Doc. TODO Sven
      */
-    public Building updateBuilding(int id, Building newBuilding) throws BuildingNotFoundException {
+    public Building updateBuilding(Long id, Building newBuilding) throws BuildingNotFoundException {
         return buildingRepository.findById(id)
             .map(building -> {
                 newBuilding.setId(id);
