+++ server/src/main/java/nl/tudelft/oopp/group39/room/exceptions/RoomNotFoundException.java
@@ -1,7 +1,7 @@
 package nl.tudelft.oopp.group39.room.exceptions;
 
 public class RoomNotFoundException extends RuntimeException {
-    public RoomNotFoundException(int id) {
+    public RoomNotFoundException(long id) {
         super("Room with id " + id + " wasn't found.");
     }
